+++ crates/apollo-compiler/tests/validation/variable.rs
@@ -355,13 +355,6 @@ fn variables_in_const_contexts() {
     let errors = schema.validate().unwrap_err().errors;
     let expected = expect_test::expect![[r#"
         Error: variable `$x` is not defined
-           ╭─[input.graphql:3:34]
-           │
-         3 │             arg: InputObj = {x: ["x"]} @dir2(arg: "x")
-           │                                  ─┬─  
-           │                                   ╰─── not found in this scope
-        ───╯
-        Error: variable `$x` is not defined
            ╭─[input.graphql:3:51]
            │
          3 │             arg: InputObj = {x: ["x"]} @dir2(arg: "x")
@@ -404,13 +397,6 @@ fn variables_in_const_contexts() {
             │                                                     ╰─── not found in this scope
         ────╯
         Error: variable `$x` is not defined
-            ╭─[input.graphql:43:39]
-            │
-         43 │                 arg2: InputObj = {x: ["x"]} @dir(arg: {x: ["x"]})
-            │                                       ─┬─  
-            │                                        ╰─── not found in this scope
-        ────╯
-        Error: variable `$x` is not defined
             ╭─[input.graphql:43:60]
             │
          43 │                 arg2: InputObj = {x: ["x"]} @dir(arg: {x: ["x"]})
@@ -437,13 +423,6 @@ fn variables_in_const_contexts() {
          48 │         interface Inter @dir(arg: {x: ["x"]}) {
             │                                        ─┬─  
             │                                         ╰─── not found in this scope
-        ────╯
-        Error: variable `$x` is not defined
-            ╭─[input.graphql:51:39]
-            │
-         51 │                 arg2: InputObj = {x: ["x"]} @dir(arg: {x: ["x"]})
-            │                                       ─┬─  
-            │                                        ╰─── not found in this scope
         ────╯
         Error: variable `$x` is not defined
             ╭─[input.graphql:51:60]
@@ -516,13 +495,6 @@ fn variables_in_const_contexts() {
             │                                    ╰─── not found in this scope
         ────╯
         Error: variable `$x` is not defined
-            ╭─[input.graphql:66:28]
-            │
-         66 │             x: [String] = ["x"] @dir2(arg: "x")
-            │                            ─┬─  
-            │                             ╰─── not found in this scope
-        ────╯
-        Error: variable `$x` is not defined
             ╭─[input.graphql:66:44]
             │
          66 │             x: [String] = ["x"] @dir2(arg: "x")
@@ -538,10 +510,19 @@ fn variables_in_const_contexts() {
         ────╯
     "#]];
     expected.assert_eq(&errors.to_string());
-    let expected_schema_errors = 26;
+    let expected_schema_errors = 22;
     assert_eq!(errors.len(), expected_schema_errors);
+
+    // Default values not validated yet: https://github.com/apollographql/apollo-rs/issues/928
+    // * @dir(arg:)
+    // * Query.field(arg2:)
+    // * Inter.field(arg2:)
+    // * InputObj.x
+    let input_default_values_not_yet_validated = 4;
     assert_eq!(
         input.matches("\"x\"").count(),
-        expected_schema_errors + expected_executable_errors
+        expected_schema_errors
+            + expected_executable_errors
+            + input_default_values_not_yet_validated
     )
 }